requirements.txt
#easysnmp
Pillow
pyusb
zeroconf
